+++ app/tasks/report/ReportRemindersTask.scala
@@ -12,8 +12,8 @@ import play.api.Logger
 import repositories.event.EventRepositoryInterface
 import repositories.report.ReportRepositoryInterface
 import repositories.tasklock.TaskRepositoryInterface
-import services.emails.Email.ProReportsUnreadReminder
 import services.emails.EmailDefinitionsPro.ProReportsReadReminder
+import services.emails.EmailDefinitionsPro.ProReportsUnreadReminder
 import services.emails.Email
 import services.emails.MailServiceInterface
 import tasks.ScheduledTask
@@ -102,7 +102,7 @@ class ReportRemindersTask(
               notReadByProsSent <- sendReminderEmailIfAtLeastOneReport(
                 notReadByPros,
                 users,
-                ProReportsUnreadReminder,
+                ProReportsUnreadReminder.EmailImpl,
                 EMAIL_PRO_REMIND_NO_READING
               )
             } yield List(readByProsSent, notReadByProsSent).flatten
